fixes for arya to be ready to run
